+++ eggs/naaya.content.talkback/naaya/content/talkback/permissions.py
@@ -6,6 +6,7 @@ PERMISSION_REVIEW_TALKBACKCONSULTATION = 'Naaya - Review TalkBack Consultation'
 PERMISSION_REVIEW_TALKBACKCONSULTATION_AFTER_DEADLINE = \
     'Naaya - Review TalkBack after deadline'
 PERMISSION_MANAGE_TALKBACKCONSULTATION = 'Naaya - Manage TalkBack Consultation'
+PERMISSION_MANAGE_COMMENTS = 'Naaya - Manage TalkBack Comments'
 PERMISSION_INVITE_TO_TALKBACKCONSULTATION = \
     'Naaya - Invite to TalkBack Consultation'
 
@@ -36,6 +37,12 @@ permission_data = {
     PERMISSION_MANAGE_TALKBACKCONSULTATION: {
         'title': 'TalkBack Consultation - manage',
         'description': """
+        """,
+    },
+    PERMISSION_MANAGE_COMMENTS: {
+        'title': 'TalkBack Consultation - manage comments',
+        'description': """
+            Permission to manage comments inside the talkback consultation
         """,
     },
 }